Refactored live search components
