+++ client/src/dropdown-items/CountryDropdownItem.js
@@ -3,7 +3,7 @@ import React from 'react';
 
 import DropdownItem from './DropdownItem';
 
-import { countryShape } from '../api';
+import { countryShape } from '../api/shapes';
 
 const CountryDropdownItem = ({ index, result }) => (
   <DropdownItem eventKey={index}>
